// src/server/api/models/investing/models/InvestmentAccountLien.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * A JSON blob representing an account lien. The fields marked with an * means that they are required.
 */
export type InvestmentAccountLien = {
    /**
     * The unique ID for the record. This is the ID that is used in all services and related records when referencing the record.
     */
    id: string;
    /**
     * The ID of the user that created the record.
     */
    createdBy: string;
    /**
     * Indicates when the record was created.
     */
    createdAt: string;
    /**
     * The ID of the user that last modified the record.
     */
    updatedBy: string;
    /**
     * Indicates when the record was last modified.
     */
    updatedAt: string;
    /**
     * Indicates if this is a deleted record
     */
    deleted: boolean;
    /**
     * A unique identifier for the client that is generated by the crm service and is required when creating a new account.
     */
    clientId: string;
    /**
     * A unique client code that is generated by the crm service and is populated by the investing service.
     */
    clientCode?: string;
    /**
     * A label that describes the client and is populated by the investing service.
     */
    clientLabel: string;
    /**
     * The account type. See the enumeration for acceptable values.
     */
    status: InvestmentAccountLien.status;
    /**
     * A unique code that is generated by the investing service.
     */
    lienNo: string;
    /**
     * A text value that describes the lien
     */
    lienLabel: string;
    /**
     * Any notes that apply to this lien
     */
    notes?: string;
    /**
     * Specify the lien amount
     */
    amount: number;
    /**
     * The lien currency.
     */
    currency: string;
    /**
     * The date that the lien should expire
     */
    expirationDate: string;
    /**
     * A unique investment account code that is generated by the investing service.
     */
    accountNo: string;
    /**
     * A unique identifier for the investment account that is generated by the investing service.
     */
    accountId: string;
    /**
     * A label that describes the investment account and is populated by the investing service when validating an order.
     */
    accountLabel: string;
};

export namespace InvestmentAccountLien {

    /**
     * The account type. See the enumeration for acceptable values.
     */
    export enum status {
        PENDING = 'PENDING',
        ACTIVE = 'ACTIVE',
        CLOSED = 'CLOSED',
    }


}